Fix: reading values when multiple services/characteristics used with same UUID
Earlier when discovering services/characteristics UUID comparison was used to
determine for a given service/characteristics if CBPeripheralDelegate callback
was related to a given service/characteristic instance.
It worked only when each service/characteristic had unique UUID.
This approach failed to work when multiple service/characteristic instances used same UUID.
The other issue was that service/characteristics discovery assumed that the result list count
will be the same as the request UUID list count. In practice the result list can be bigger than
the request UUID list because of the possibility of redundant UUIDs.

The commit fixes both issues above:

- Refactored RxServiceType, RxCharacteristicType, RxDescriptorType equality.
Service/Characteristic/Descriptor instances will be considered equal if the the wrapped
CoreBluetooth objects references are equal. This guarantees proper bindings with
CBPeripheralDelegate callbacks with corresponding CoreBluetooth objects.

- Result list is filtered based on the requested UUID list

- Some code in unit tests is updated to make sure unit tests are passing